Move TPCH-related tests to tpch package
The package already existed.
