+++ testing/trino-tests/src/test/java/io/trino/tests/tpch/TestTpchDistributedStats.java
@@ -11,13 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.tests;
+package io.trino.tests.tpch;
 
 import com.google.common.collect.ImmutableMap;
 import io.trino.plugin.tpch.ColumnNaming;
 import io.trino.testing.DistributedQueryRunner;
 import io.trino.testing.statistics.StatisticsAssertion;
-import io.trino.tests.tpch.TpchQueryRunnerBuilder;
 import io.trino.tpch.TpchTable;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
