+++ testing/trino-tests/src/test/java/io/trino/tests/tpch/TestTpchLocalStats.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.tests;
+package io.trino.tests.tpch;
 
 import com.google.common.collect.ImmutableMap;
 import io.trino.Session;
